adapting the test suite
