+++ acceptance_tests/junitxml/results.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?><testsuite errors="0" failures="0" name="pytest" skips="0" tests="3" time="39.389"><testcase classname="tests.test_wfs" file="tests/test_wfs.py" line="1" name="test_get_feature" time="35.91578412055969"><system-out>Removing geoserver_geoserver_1 ... 
+Removing geoserver_geoserver_1 ... done
+Going to remove geoserver_geoserver_1
+geoserver uses an image, skipping
+Creating geoserver_geoserver_1 ... 
+Creating geoserver_geoserver_1 ... done
+TEST                 | 2021-12-16 15:20:41,828  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:43,834  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:45,838  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:47,843  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:49,849  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:51,853  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:53,858  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:55,861  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:57,866  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:20:59,871  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:21:01,875  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:21:03,882  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:21:05,887  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:21:07,891  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:21:09,897  INFO tests.conftest Trying to connect to GeoServer... 
+TEST                 | 2021-12-16 15:21:10,292  INFO tests.conftest GeoServer started
+</system-out></testcase><testcase classname="tests.test_wms" file="tests/test_wms.py" line="0" name="test_get_capabilities" time="0.8213269710540771"></testcase><testcase classname="tests.test_wms" file="tests/test_wms.py" line="6" name="test_get_map" time="2.637129783630371"></testcase></testsuite>
