+++ acceptance_tests/tests/__init__.py
@@ -1,0 +1,12 @@
+# import logging
+#
+# import netifaces
+# import sys
+#
+# from acceptance_tests import utils
+
+# import os
+#
+#
+# def in_docker():
+#     return os.environ.get("DOCKER_RUN", "0") == "1"
